moved all the components over to the new i18n format and setup color config files
