+++ home_intent/default_configs/de/home_assistant/color_names.yaml
@@ -1,0 +1,143 @@
+color_names:
+  Alte Spitze: oldlace
+  Altrosa: mistyrose
+  Antikweiß: antiquewhite
+  Aquamarinblau: aquamarine
+  Backstein: firebrick
+  Beige: beige
+  Biskuit: bisque
+  Blasse Goldrutenfarbe: palegoldenrod
+  Blasses Violettrot: palevioletred
+  Blassgrün: palegreen
+  Blasstürkis: paleturquoise
+  Blau: blue
+  Blauviolett: blueviolet
+  Blütenweiß: floralwhite
+  Braun: brown
+  Chiffongelb: lemonchiffon
+  Cremige Minze: mintcream
+  Distel: thistle
+  Dunkelblau: darkblue
+  Dunkelgrau: darkgrey
+  Dunkelgrau: dimgrey
+  Dunkelgrün: darkgreen
+  Dunkelkhaki: darkkhaki
+  Dunkelmagenta: darkmagenta
+  Dunkelrot: darkred
+  Dunkeltürkis: darkcyan
+  Dunkeltürkis: darkturquoise
+  Dunkelvilolett: darkviolet
+  Dunkle Goldrutenfarbe: darkgoldenrod
+  Dunkle Lachsfarbe: darksalmon
+  Dunkle Orchidee: darkorchid
+  Dunkles Olivgrün: darkolivegreen
+  Dunkles Orange: darkorange
+  Dunkles Schieferblau: darkslateblue
+  Dunkles Schiefergrau: darkslateyellow
+  Dunkles Seegrün: darkseagreen
+  Eisfarben: aliceblue
+  Elfenbein: ivory
+  Frühlingsgrün: springgreen
+  Fuchsia: fuchsia
+  Gainsboro: gainsboro
+  Geisterweiß: ghostwhite
+  Gelb: yellow
+  Gelbbraun: burlywood
+  Gelbgrün: yellowgreen
+  Gold: gold
+  Goldrute: goldenrod
+  Grau: gray
+  Grau: grey
+  Grün: green
+  Grüngelb: greenyellow
+  Hautfarben: tan
+  Hellblau: lightblue
+  Helle Lachsfarbe: lightsalmon
+  Helles Cyan: lightcyan
+  Helles Goldrutengelb: lightgoldenrodyellow
+  Helles Himmelblau: lightskyblue
+  Helles Korallenrot: lightcoral
+  Helles Schiefergrau: lightslategrey
+  Helles Seegrün: lightseagreen
+  Helles Stahlblau: lightsteelblue
+  Hellgelb: lightyellow
+  Hellgrau: lightgray
+  Hellgrau: lightgrey
+  Hellgrün: chartreuse
+  Hellgrün: lightgreen
+  Hellrosa: lightpink
+  Himmelblau: azure
+  Himmelblau: skyblue
+  Honigmelone: honeydew
+  Indigo: indigo
+  Indischrot: indianred
+  Kadettenblau: cadetblue
+  Karmesinrot: crimson
+  Kastanie: maroon
+  Koralle: coral
+  Kornblumenblau: cornflowerblue
+  Krickentengrün: teal
+  Königsbalu: royalblue
+  Lachsfarben: salmon
+  Lavendel: lavender
+  Lavendelrosa: lavenderblush
+  Leinen: linen
+  Leuchtendes Rosa: hotpink
+  Limone: lime
+  Limonengrün: limegreen
+  Magenta: magenta
+  Mais: cornsilk
+  Mandelweiß: blanchedalmond
+  Marinenblau: navy
+  Mitternachtsblau: midnightblue
+  Mittlere Orchidee: mediumorchid
+  Mittlere Türkis: mediumturquoise
+  Mittleres Aquamarin: mediumaquamarine
+  Mittleres Blau: mediumblue
+  Mittleres Frühlingsgrün: mediumspringgreen
+  Mittleres Schieferblau: mediumslateblue
+  Mittleres Seegrün: mediumseagreen
+  Mittleres Violett: mediumpurple
+  Mittleres Violettrot: mediumvioletred
+  Mokassin: moccasin
+  Muschel: seashell
+  Navajoweiß: navajowhite
+  Olivgraubraun: olivedrab
+  Olivgrün: olive
+  Orange: orange
+  Orangerot: orangered
+  Orchidee: orchid
+  Papayacreme: papayawhip
+  Persenningblau: dodgerblue
+  Peru: peru
+  Pfirsich: peachpuff
+  Pflaume: plum
+  Purpur: purple
+  Rasengrün: lawngreen
+  Rauchfarben: whitesmoke
+  Rosa: pink
+  Rosiges Braun: rosybrown
+  Rot: red
+  Sandbraun: sandybrown
+  Sattelbraun: saddlebrown
+  Schieferblau: slateblue
+  Schiefergrau: slategrey
+  Schneeweiß: snow
+  Schokolade: chocolate
+  Schwarz: black
+  Seegrün: seagreen
+  Siennaerde: sienna
+  Silber: silver
+  Stahlblau: steelblue
+  Staubfarben: khaki
+  Taubenblau: powderblue
+  Tiefes Himmelblau: deepskyblue
+  Tiefrosa: deeppink
+  Tomate: tomato
+  Türkis: cyan
+  Türkis: turquoise
+  Veilchen: violet
+  Waldgrün: forestgreen
+  Wasser: aqua
+  Weizen: wheat
+  Weiß: white
