+++ home_intent/default_configs/en/home_assistant/color_names.yaml
@@ -1,0 +1,143 @@
+color_names:
+  Alice Blue: aliceblue
+  Antique White: antiquewhite
+  Aqua: aqua
+  Aqua marine: aquamarine
+  Azure: azure
+  Beige: beige
+  Bisque: bisque
+  Black: black
+  Blanched almond: blanchedalmond
+  Blue: blue
+  Blue violet: blueviolet
+  Brown: brown
+  Burly wood: burlywood
+  Cadet Blue: cadetblue
+  Chartreuse: chartreuse
+  Chocolate: chocolate
+  Coral: coral
+  Corn flower Blue: cornflowerblue
+  Corn silk: cornsilk
+  Crimson: crimson
+  Cyan: cyan
+  Dark Blue: darkblue
+  Dark Cyan: darkcyan
+  Dark Golden rod: darkgoldenrod
+  Dark Green: darkgreen
+  Dark Grey: darkgrey
+  Dark Khaki: darkkhaki
+  Dark Magenta: darkmagenta
+  Dark Olive Green: darkolivegreen
+  Dark Orange: darkorange
+  Dark Orchid: darkorchid
+  Dark Red: darkred
+  Dark Salmon: darksalmon
+  Dark Sea Green: darkseagreen
+  Dark Slate Blue: darkslateblue
+  Dark Slate Gray: darkslategray
+  Dark Turquoise: darkturquoise
+  Dark Violet: darkviolet
+  Deep Pink: deeppink
+  Deep Sky Blue: deepskyblue
+  Dim Grey: dimgrey
+  Dodger Blue: dodgerblue
+  Firebrick: firebrick
+  Floral White: floralwhite
+  Forest Green: forestgreen
+  Fuchsia: fuchsia
+  Gainsboro: gainsboro
+  Ghost White: ghostwhite
+  Gold: gold
+  Golden rod: goldenrod
+  Gray: gray
+  Green: green
+  Green Yellow: greenyellow
+  Grey: grey
+  Honey dew: honeydew
+  Hot Pink: hotpink
+  Indian Red: indianred
+  Indigo: indigo
+  Ivory: ivory
+  Khaki: khaki
+  Lavender: lavender
+  Lavender Blush: lavenderblush
+  Lawn Green: lawngreen
+  Lemon chiffon: lemonchiffon
+  Light Goldenrod Yellow: lightgoldenrodyellow
+  Light Blue: lightblue
+  Light Coral: lightcoral
+  Light Cyan: lightcyan
+  Light Gray: lightgray
+  Light Green: lightgreen
+  Light Grey: lightgrey
+  Light Pink: lightpink
+  Light Salmon: lightsalmon
+  Light Sea Green: lightseagreen
+  Light Sky Blue: lightskyblue
+  Light Slate Grey: lightslategrey
+  Light Steel Blue: lightsteelblue
+  Light Yellow: lightyellow
+  Lime: lime
+  Lime Green: limegreen
+  Linen: linen
+  Magenta: magenta
+  Maroon: maroon
+  Medium Aquamarine: mediumaquamarine
+  Medium Blue: mediumblue
+  Medium Orchid: mediumorchid
+  Medium Purple: mediumpurple
+  Medium Sea Green: mediumseagreen
+  Medium Slate Blue: mediumslateblue
+  Medium Spring Green: mediumspringgreen
+  Medium Turquoise: mediumturquoise
+  Medium Violet Red: mediumvioletred
+  Midnight Blue: midnightblue
+  Mint cream: mintcream
+  Misty rose: mistyrose
+  Moccasin: moccasin
+  Navajo White: navajowhite
+  Navy: navy
+  Old lace: oldlace
+  Olive: olive
+  Olive drab: olivedrab
+  Orange: orange
+  Orange Red: orangered
+  Orchid: orchid
+  Pale Golden rod: palegoldenrod
+  Pale Green: palegreen
+  Pale Turquoise: paleturquoise
+  Pale Violet Red: palevioletred
+  Papaya whip: papayawhip
+  Peach puff: peachpuff
+  Peru: peru
+  Pink: pink
+  Plum: plum
+  Powder Blue: powderblue
+  Purple: purple
+  Red: red
+  Rosy Brown: rosybrown
+  Royal Blue: royalblue
+  Saddle Brown: saddlebrown
+  Salmon: salmon
+  Sandy Brown: sandybrown
+  Sea Green: seagreen
+  Seashell: seashell
+  Sienna: sienna
+  Silver: silver
+  Sky Blue: skyblue
+  Slate Blue: slateblue
+  Slate Grey: slategrey
+  Snow: snow
+  Spring Green: springgreen
+  Steel Blue: steelblue
+  Tan: tan
+  Teal: teal
+  Thistle: thistle
+  Tomato: tomato
+  Turquoise: turquoise
+  Violet: violet
+  Wheat: wheat
+  White: white
+  White Smoke: whitesmoke
+  Yellow: yellow
+  Yellow Green: yellowgreen
