+++ home_intent/default_configs/es/home_assistant/color_names.yaml
@@ -1,0 +1,143 @@
+color_names:
+  Agua: aqua
+  Aguamarina Mediana: mediumaquamarine
+  Aguamarina: aquamarine
+  Almendra Pálido: blanchedalmond
+  Amarillo Claro: lightyellow
+  Amarillo Dorado Claro: lightgoldenrodyellow
+  Amarillo: yellow
+  Ambrosía: honeydew
+  Azul Acero Brillante: lightsteelblue
+  Azul Aciano: cornflowerblue
+  Azul Alicia: aliceblue
+  Azul Cadete: cadetblue
+  Azul Cielo: azure
+  Azul Luz: lightblue
+  Azul Marino: navy
+  Azul Medianoche: midnightblue
+  Azul Medio: mediumblue
+  Azul Oscuro: darkblue
+  Azul Palomas: powderblue
+  Azul Pizarra Mediana: mediumslateblue
+  Azul Pizarra Oscuro: darkslateblue
+  Azul Pizarra: slateblue
+  Azul Real: royalblue
+  Azul de Acero: steelblue
+  Azul: blue
+  Beige: beige
+  Bisque: bisque
+  Blanco Antigua: antiquewhite
+  Blanco Floral: floralwhite
+  Blanco Humo: whitesmoke
+  Blanco Navajo: navajowhite
+  Blanco Nieve: snow
+  Blanco: white
+  Bronceado: tan
+  Caqui Oscuro: darkkhaki
+  Caqui: khaki
+  Cardo: thistle
+  Carmesi: crimson
+  Chartreuse: chartreuse
+  Chocolate: chocolate
+  Cian Luz: lightcyan
+  Cian: cyan
+  Cielo Azul Brillante: lightskyblue
+  Cielo Azul Profundo: deepskyblue
+  Cielo Azul: skyblue
+  Ciruela: plum
+  Concha: seashell
+  Coral: coral
+  Crema de Papaya: papayawhip
+  Cubierta Azul: dodgerblue
+  Encaje Viejo: oldlace
+  Fantasma Blanco: ghostwhite
+  Fucsia: fuchsia
+  Gainsboro: gainsboro
+  Gasa Amarilla: lemonchiffon
+  Granate: maroon
+  Gris Claro: lightgray
+  Gris Claro: lightgrey
+  Gris Oliva Marrón: olivedrab
+  Gris Oscuro: darkgrey
+  Gris Oscuro: dimgrey
+  Gris Pizarra Brillante: lightslategrey
+  Gris Pizarra Oscuro: darkslateyellow
+  Gris Pizarra: slategrey
+  Gris: gray
+  Gris: grey
+  Ladrillo Refractario: firebrick
+  Lavanda: lavender
+  Lima: lime
+  Lino: linen
+  Madera: burlywood
+  Magenta Oscura: darkmagenta
+  Magenta: magenta
+  Mar Verde Medio: mediumseagreen
+  Marfil: ivory
+  Marrón Arena: sandybrown
+  Marrón Ensillar: saddlebrown
+  Marrón Rosado: rosybrown
+  Marrón: brown
+  Maíz: cornsilk
+  Melocotón: peachpuff
+  Menta Cremosa: mintcream
+  Mocassin: moccasin
+  Naranja Oscura: darkorange
+  Naranja: orange
+  Negro: black
+  Oro: gold
+  Orquidea Media: mediumorchid
+  Orquidea Oscura: darkorchid
+  Orquídea: orchid
+  Oscuro mar Verde: darkseagreen
+  Perú: peru
+  Plata: silver
+  Púrpura Medio: mediumpurple
+  Púrpura Oscuro: darkviolet
+  Púrpura: purple
+  Rojo Anaranjado: orangered
+  Rojo Coral Brillante: lightcoral
+  Rojo Indio: indianred
+  Rojo Oscura: darkred
+  Rojo Púrpura Medio: mediumvioletred
+  Rojo Viejo: mistyrose
+  Rojo Violeta Pálido: palevioletred
+  Rojo: red
+  Rosa Brillante: hotpink
+  Rosa Claro: lightpink
+  Rosa Lavanda: lavenderblush
+  Rosa Profundo: deeppink
+  Rosa: pink
+  Salmón Claro: lightsalmon
+  Salmón Oscuro: darksalmon
+  Salmón: salmon
+  Solidago: goldenrod
+  Tierra de Siena: sienna
+  Tomate: tomato
+  Trigo: wheat
+  Turquesa Medio: mediumturquoise
+  Turquesa Oscuro: darkcyan
+  Turquesa Oscuro: darkturquoise
+  Turquesa Pálido: paleturquoise
+  Turquesa: turquoise
+  Vara de Oro Oscura: darkgoldenrod
+  Vara de Oro Pálida: palegoldenrod
+  Verde Amarillo: greenyellow
+  Verde Amarillo: yellowgreen
+  Verde Azulado: teal
+  Verde Bosque: forestgreen
+  Verde Claro: lightgreen
+  Verde Césped: lawngreen
+  Verde Lago: seagreen
+  Verde Lima: limegreen
+  Verde Mar Claro: lightseagreen
+  Verde Oliva Oscuro: darkolivegreen
+  Verde Oliva: olive
+  Verde Oscuro: darkgreen
+  Verde Primavera Mediana: mediumspringgreen
+  Verde Pálido: palegreen
+  Verde del Resorte: springgreen
+  Verde: green
+  Violeta Azul: blueviolet
+  Violeta: violet
+  Índigo: indigo
