+++ home_intent/default_configs/fr/home_assistant/color_names.yaml
@@ -1,0 +1,143 @@
+color_names:
+  Aigue-Marine Moyenne: mediumaquamarine
+  Aigue-Marine: aquamarine
+  Amande Pâle: blanchedalmond
+  Ambroisie: honeydew
+  Aqua: aqua
+  Argent: silver
+  Beige: beige
+  Bisque: bisque
+  Blanc Fantôme: ghostwhite
+  Blanc Floral: floralwhite
+  Blanc Navajo: navajowhite
+  Blanc Neige: snow
+  Blanc: white
+  Blanche Fumée: whitesmoke
+  Bleu Acier Brillant: lightsteelblue
+  Bleu Acier: steelblue
+  Bleu Alicia: aliceblue
+  Bleu Ardoise Foncé: darkslateblue
+  Bleu Ardoise Moyenne: mediumslateblue
+  Bleu Ardoise: slateblue
+  Bleu Cadet: cadetblue
+  Bleu Ciel Profond: deepskyblue
+  Bleu Ciel: azure
+  Bleu Ciel: skyblue
+  Bleu Foncé: darkblue
+  Bleu Marine: navy
+  Bleu Moyen: mediumblue
+  Bleu Royal: royalblue
+  Bleu Violet: blueviolet
+  Bleu de Pigeons: powderblue
+  Bleu: blue
+  Bleue Lumière: lightblue
+  Bleue Minuit: midnightblue
+  Bleuet: cornflowerblue
+  Blé: wheat
+  Bois: burlywood
+  Brique Réfractaire: firebrick
+  Brun Rosé: rosybrown
+  Brun Sable: sandybrown
+  Brune Selle: saddlebrown
+  Chardon: thistle
+  Chartreuse: chartreuse
+  Chocolat: chocolate
+  Ciel Bleu Clair: lightskyblue
+  Citron: lime
+  Coquillage: seashell
+  Corail: coral
+  Corn: cornsilk
+  Couverture Bleu: dodgerblue
+  Cramoisi: crimson
+  Cris Clair: lightgrey
+  Crème de Papaye: papayawhip
+  Cyan Clair: lightcyan
+  Cyan: cyan
+  Fuchsia: fuchsia
+  Gainsboro: gainsboro
+  Gris Ardoise Brillant: lightslategrey
+  Gris Ardoise Foncé: darkslateyellow
+  Gris Ardoise: slategrey
+  Gris Clair: lightgray
+  Gris Foncé: darkgrey
+  Gris Foncé: dimgrey
+  Gris: gray
+  Gris: grey
+  Indigo: indigo
+  Ivoire: ivory
+  Jaune Clair: lightgoldenrodyellow
+  Jaune Clair: lightyellow
+  Jaune Olive: olive
+  Jaune: yellow
+  Kaki Foncé: darkkhaki
+  Kaki: khaki
+  Lavande: lavender
+  Lin: linen
+  Magenta Foncé: darkmagenta
+  Magenta: magenta
+  Marron / Bordeaux: maroon
+  Marron: brown
+  Menthe Crémeuse: mintcream
+  Mer Vert foncé: darkseagreen
+  Mocassin: moccasin
+  Mousseline Jaune: lemonchiffon
+  Noir: black
+  Olive Gris Brun: olivedrab
+  Or: gold
+  Orange Foncé: darkorange
+  Orange: orange
+  Orchidée Moyenne: mediumorchid
+  Orchidée Noire: darkorchid
+  Orchidée: orchid
+  Pourpre / Violet: purple
+  Prune: plum
+  Pérou: peru
+  Pêche: peachpuff
+  Rose Brillant: hotpink
+  Rose Clair: lightpink
+  Rose Lavande: lavenderblush
+  Rose Profond: deeppink
+  Rose: pink
+  Rouge Corail Vif: lightcoral
+  Rouge Indien: indianred
+  Rouge Noire: darkred
+  Rouge Orange: orangered
+  Rouge Pourpre Moyen: mediumvioletred
+  Rouge: red
+  Sarcelle: teal
+  Saumon Clair: lightsalmon
+  Saumon Foncé: darksalmon
+  Saumon: salmon
+  Solidago: goldenrod
+  Tanné: tan
+  Terre de Sienne: sienna
+  Tomate: tomato
+  Turquoise Foncé: darkcyan
+  Turquoise Foncé: darkturquoise
+  Turquoise Moyen: mediumturquoise
+  Turquoise Pâle: paleturquoise
+  Turquoise: turquoise
+  Verge d'Or Pâle: palegoldenrod
+  Verge d'or Sombre: darkgoldenrod
+  Vert Citron: limegreen
+  Vert Clair: lightgreen
+  Vert Foncé: darkgreen
+  Vert Forêt: forestgreen
+  Vert Jaune: greenyellow
+  Vert Jaune: yellowgreen
+  Vert Lac: seagreen
+  Vert Mer Clair: lightseagreen
+  Vert Olive Foncé: darkolivegreen
+  Vert Pelouse: lawngreen
+  Vert Printemps: springgreen
+  Vert Pâle: palegreen
+  Vert de Mer Moyen: mediumseagreen
+  Vert de Printemps Moyen: mediumspringgreen
+  Vert: green
+  Vieille Dentelle: oldlace
+  Vieux Blanc: antiquewhite
+  Vieux Rose: mistyrose
+  Violet Foncé: darkviolet
+  Violet Moyen: mediumpurple
+  Violet Rouge Pâle: palevioletred
+  Violet: violet
