--- home_intent/default_configs/home_assistant/colors.txt
@@ -1,148 +1,0 @@
-alice blue
-antique white
-aquamarine
-aqua
-azure
-beige
-bisque
-black
-blanched almond
-blue violet
-blue
-brown
-burly wood
-cadet blue
-chartreuse
-chocolate
-coral
-cornflower blue
-corn silk
-crimson
-cyan
-dark blue
-dark cyan
-dark goldenrod
-dark gray
-dark green
-dark grey
-dark khaki
-dark magenta
-dark olive green
-dark orange
-dark orchid
-dark red
-dark salmon
-dark sea green
-dark slate blue
-dark slate gray
-dark slate grey
-dark turquoise
-dark violet
-deep pink
-deep sky blue
-dim gray
-dim grey
-dodger blue
-firebrick
-floral white
-forest green
-fuchsia
-gainsboro
-ghost white
-goldenrod
-gold
-gray
-green yellow
-green
-grey
-home assistant
-honeydew
-hot pink
-indian red
-indigo
-ivory
-khaki
-lavender blush
-lavender
-lawn green
-lemon chiffon
-light blue
-light coral
-light cyan
-light goldenrod yellow
-light gray
-light green
-light grey
-light pink
-light salmon
-light sea green
-light sky blue
-light slate gray
-light slate grey
-light steel blue
-light yellow
-lime green
-linen
-magenta
-maroon
-medium aquamarine
-medium blue
-medium orchid
-medium purple
-medium sea green
-medium slate blue
-medium spring green
-medium turquoise
-medium violet red
-midnight blue
-mint cream
-misty rose
-moccasin
-navajo white
-navy blue
-navy
-old lace
-olive drab
-olive
-orange red
-orange
-orchid
-pale goldenrod
-pale green
-pale turquoise
-pale violet red
-papaya whip
-peach puff
-peru
-pink
-plum
-powder blue
-purple
-red
-rosy brown
-royal blue
-saddle brown
-salmon
-sandy brown
-sea green
-seashell
-sienna
-silver
-sky blue
-slate blue
-slate gray
-slate grey
-snow
-spring green
-steel blue
-tan
-teal
-thistle
-tomato
-turquoise
-violet
-wheat
-white smoke
-white
-yellow green
-yellow
